digital signals cleanup

+++ implot.h
@@ -88,7 +88,7 @@ enum ImPlotStyleVar_ {
     ImPlotStyleVar_MarkerWeight,     // float, outline weight of markers in pixels
     ImPlotStyleVar_ErrorBarSize,     // float, error bar whisker width in pixels
     ImPlotStyleVar_ErrorBarWeight,   // float, error bar whisker weight in pixels
-    ImPlotStyleVar_DigitalBitHeight, // int, digital channels bit height (at 1)
+    ImPlotStyleVar_DigitalBitHeight, // float, digital channels bit height (at 1)
     ImPlotStyleVar_COUNT
 };
 
@@ -122,7 +122,7 @@ struct ImPlotStyle {
     float    MarkerWeight;            // = 1, outline weight of markers in pixels
     float    ErrorBarSize;            // = 5, error bar whisker width in pixels
     float    ErrorBarWeight;          // = 1.5, error bar whisker weight in pixels
-    int      DigitalBitHeight;        // = 7, digital channels bit height (at 1)
+    float    DigitalBitHeight;        // = 8, digital channels bit height (at y = 1.0f)
     ImVec4   Colors[ImPlotCol_COUNT]; // array of plot specific colors
     ImPlotStyle();
 };
